amend

+++ Code/Core/Objectiks/Models/DocumentStorage.cs
@@ -94,7 +94,10 @@ namespace Objectiks.Models
                         File.Delete(Backup);
                     }
                     catch { }
+                }
 
+                if (File.Exists(Target))
+                {
                     File.Copy(Target, Backup);
                 }
